requirements.txt
posterioragreement @ git+https://github.com/viictorjimenezzz/posterioragreement.git@main
diagvibsix @ git+https://github.com/viictorjimenezzz/diagvibsix.git@59cb3760c097a1068621c015e688543d2c854f98
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.6.0
ansible-cmdb==1.31
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
autoattack==0.1
autocommand==2.2.2
autopage==0.5.1
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.2.0
cheroot==10.0.0
CherryPy==18.9.0
click==8.1.6
cliff==4.3.0
cloudpickle==2.2.1
cmaes==0.10.0
cmd2==2.4.3
colorlog==6.7.0
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
docker-pycreds==0.4.0
eagerpy==0.30.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.12.2
fonttools==4.41.0
foolbox==3.3.3
frozenlist==1.4.0
fsspec==2023.6.0
geotorch==0.3.0
gevent==23.7.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
huggingface-hub==0.21.4
hydra-colorlog==1.2.0
hydra-core==1.3.1
hydra-optuna-sweeper==1.2.0
hydra-submitit-launcher==1.2.0
idna==2.10
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflect==7.0.0
ipdb==0.13.13
ipython==8.18.1
jaraco.collections==5.0.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.1
jsonxs==0.6
kiwisolver==1.4.4
lazy_loader==0.3
lightning-fabric==1.9.1
lightning-utilities==0.9.0
littleutils==0.2.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.4
networkx==3.1
numpy==1.22.4
ogb==1.3.6
omegaconf==2.3.0
optuna==2.10.1
outdated==0.2.2
packaging==23.1
pandas==1.3.5
parso==0.8.3
pathtools==0.1.2
pbr==5.11.1
pexpect==4.9.0
Pillow==10.0.0
portend==3.2.0
prettytable==3.8.0
prompt-toolkit==3.0.43
protobuf==3.20.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.15.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrootutils==1.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-lightning==1.9.1
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.25.1
rich==13.4.2
robustbench==1.1
safetensors==0.4.2
SciencePlots==2.1.0
scikit-image==0.22.0
scikit-learn==1.1.3
scipy==1.11.1
seaborn==0.12.2
secml==0.15.6
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.19
stack-data==0.6.3
stevedore==5.1.0
submitit==1.4.5
tempora==5.5.1
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.8.12
timm==0.6.13
tokenizers==0.15.2
tomli==2.0.1
torch==1.10.0
torchdiffeq==0.2.3
torchmetrics==0.11.0
torchvision==0.11.1
tqdm==4.65.0
traitlets==5.14.1
transformers==4.38.2
typing==3.7.4.3
typing_extensions==4.7.1
urllib3==1.26.16
ushlex==0.99.1
wandb==0.15.5
wcwidth==0.2.6
wilds==2.0.0
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
